fix(codeql): 2 high overflow + 1 medium missing-permissions
PR #39 CodeQL flagged 3 alerts (2 high security_severity_level + 1
medium):

HIGH x2 -- cpp/integer-multiplication-cast-to-long
  r-package/morie/src/morie_dsp.cpp:111
  r-package/morie/src/morie_dsp.cpp:122

  Both sites compute `&P[r * order]` where `r` and `order` are
  both `int`. The multiplication happens in `int`, THEN the
  result is implicitly widened to `ptrdiff_t` for the pointer
  offset. With large `order` the int multiply could overflow
  before widening, producing wrong addresses (undefined behaviour).

  In practice AR `order` is small (<= ~100) so overflow is
  unreachable, but the static analyzer is right that the cast
  should happen BEFORE the multiplication. Promoted one operand
  to size_t so the multiply runs in the wider type.

MEDIUM x1 -- actions/missing-workflow-permissions
  .github/workflows/version-drift.yml

  Workflow had no top-level `permissions:` block, defaulting to
  the runner's broad GITHUB_TOKEN scope. The job only checks out
  + reads the working tree, so locked it down to:
    permissions:
      contents: read

VERSION_INVENTORY.csv regenerated (line shifts from the new
comment block in morie_dsp.cpp).

+++ r-package/morie/src/morie_dsp.cpp
@@ -108,7 +108,12 @@ List morie_dsp_rls_cpp(NumericVector x, NumericVector d,
 
         for (int r = 0; r < order; ++r) {
             double s = 0.0;
-            const double *Pr = &P[r * order];
+            // Cast to size_t BEFORE the multiplication: `int * int -> int`
+        // can overflow with large `order` (CodeQL cpp/integer-
+        // multiplication-cast-to-long flags the implicit ptrdiff_t
+        // widening). Promote one operand so the multiply happens in
+        // the wider type.
+        const double *Pr = &P[static_cast<std::size_t>(r) * static_cast<std::size_t>(order)];
             for (int c = 0; c < order; ++c) s += Pr[c] * seg[c];
             Pseg[r] = s;
         }
@@ -119,7 +124,10 @@ List morie_dsp_rls_cpp(NumericVector x, NumericVector d,
         for (int r = 0; r < order; ++r) w[r] += kg[r] * err;
         const double inv_lam = 1.0 / lam;
         for (int r = 0; r < order; ++r) {
-            double *Pr = &P[r * order];
+            // Same int*int -> ptrdiff_t widening trap as above
+            // (CodeQL cpp/integer-multiplication-cast-to-long); promote
+            // to size_t before the multiplication.
+            double *Pr = &P[static_cast<std::size_t>(r) * static_cast<std::size_t>(order)];
             const double kgr = kg[r];
             for (int c = 0; c < order; ++c) {
                 Pr[c] = (Pr[c] - kgr * Pseg[c]) * inv_lam;
